Rename db_models into models
