+++ {{cookiecutter.project_slug}}/backend/app/app/api/api_v1/endpoints/items.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 from app import crud
 from app.api.utils.db import get_db
 from app.api.utils.security import get_current_active_user
-from app.db_models.user import User as DBUser
+from app.models.user import User as DBUser
 from app.schemas.item import Item, ItemCreate, ItemUpdate
 
 router = APIRouter()
