+++ {{cookiecutter.project_slug}}/backend/app/app/api/api_v1/endpoints/users.py
@@ -9,7 +9,7 @@ from app import crud
 from app.api.utils.db import get_db
 from app.api.utils.security import get_current_active_superuser, get_current_active_user
 from app.core import config
-from app.db_models.user import User as DBUser
+from app.models.user import User as DBUser
 from app.schemas.user import User, UserCreate, UserInDB, UserUpdate
 from app.utils import send_new_account_email
