+++ {{cookiecutter.project_slug}}/backend/app/app/crud/item.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 from fastapi.encoders import jsonable_encoder
 from sqlalchemy.orm import Session
 
-from app.db_models.item import Item
+from app.models.item import Item
 from app.schemas.item import ItemCreate, ItemUpdate
 from app.crud.base import CrudBase
