+++ {{cookiecutter.project_slug}}/backend/app/app/crud/user.py
@@ -4,7 +4,7 @@ from fastapi.encoders import jsonable_encoder
 from sqlalchemy.orm import Session
 
 from app.core.security import get_password_hash, verify_password
-from app.db_models.user import User
+from app.models.user import User
 from app.schemas.user import UserCreate, UserUpdate
