+++ {{cookiecutter.project_slug}}/backend/app/app/db/base.py
@@ -1,6 +1,6 @@
 # Import all the models, so that Base has them before being
 # imported by Alembic
 from app.db.base_class import Base  # noqa
-from app.db_models.user import User  # noqa
-from app.db_models.item import Item  # noqa
-from app.db_models.subitem import SubItem  # noqa
+from app.models.user import User  # noqa
+from app.models.item import Item  # noqa
+from app.models.subitem import SubItem  # noqa
